login: refactor page to components
