+++ src/components/LoginBrandInfo/styles.module.scss
@@ -1,0 +1,46 @@
+@import 'theme/colors';
+
+.infoContainer {
+	@media (max-width: 800px) {
+		display: none;
+	}
+	color: white;
+	background-color: #181a20;
+	background: linear-gradient(130.33deg, #2b395b 0%, #212040 66.15%);
+	overflow: hidden;
+	position: relative;
+	display: flex;
+	justify-content: center;
+	align-items: flex-end;
+	padding: 2rem;
+
+	.bg {
+		position: absolute;
+		top: 0;
+		right: 0;
+		width: 100%;
+		width: 80rem;
+	}
+
+	.contentWrapper {
+		z-index: 10;
+		min-height: 10rem;
+		text-align: center;
+		margin-bottom: 5rem;
+
+		.logo {
+			width: 15rem;
+		}
+
+		p {
+			margin-top: 3rem;
+			font-size: 1.5rem;
+		}
+
+		.navigationDots img {
+			height: 0.9rem;
+			margin: 1rem 0.4rem;
+			cursor: pointer;
+		}
+	}
+}
